+++ prboom2/src/p_setup.h
@@ -46,8 +46,8 @@ void P_Init(void);               /* Called by startup code. */
 extern const byte *rejectmatrix;   /* for fast sight rejection -  cph - const* */
 
 /* killough 3/1/98: change blockmap from "short" to "long" offsets: */
-extern long     *blockmaplump;   /* offsets in blockmap are from here */
-extern long     *blockmap;
+extern int      *blockmaplump;   /* offsets in blockmap are from here */
+extern int      *blockmap;
 extern int      bmapwidth;
 extern int      bmapheight;      /* in mapblocks */
 extern fixed_t  bmaporgx;
